+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to
 
 ## [Unreleased]
 
+## 2.1.0 - 2023-05-18
+
+### Fixed
+
+- Ingested Hardware data
+
+### Added
+
+- Mapped relationship between user and hardware
+
 ## 2.0.3 - 2023-03-15
 
 ### Fixed
@@ -44,6 +54,7 @@ and this project adheres to
 | Consumable | `snipeit_consumable_resource`  | `Resource`      |
 | License    | `snipeit_licensed_application` | `Application`   |
 | User       | `snipeit_user`                 | `User`          |
+| Hardware   | `snipeit_hardware`             | `Device`        |
 
 ### Relationships
 
@@ -53,6 +64,7 @@ and this project adheres to
 | `snipeit_account`     | **HAS**               | `snipeit_licensed_application` |
 | `snipeit_account`     | **HAS**               | `snipeit_user`                 |
 | `snipeit_user`        | **USES**              | `snipeit_consumable_resource`  |
+| `snipeit_user`        | **HAS**               | `snipeit_hardware`             |
 
 ### Mapped Relationships